+++ erp/views.py
@@ -218,32 +218,6 @@ def search(request):
     )
 
 
-class CustomActivationCompleteView(TemplateView):
-    def get_context_data(self, **kwargs):
-        "Spread the next redirect value from qs param to template context key."
-        context = super().get_context_data(**kwargs)
-        context["next"] = self.request.GET.get("next", "")
-        return context
-
-
-class CustomRegistrationView(RegistrationView):
-    def get_email_context(self, activation_key):
-        "Add the next redirect value to the email template context."
-        context = super().get_email_context(activation_key)
-        context["next"] = self.request.GET.get("next", "")
-        return context
-
-
-class CustomActivationView(ActivationView):
-    def get_success_url(self, user=None):
-        "Add the next redirect path to the success redirect url"
-        url = super().get_success_url(user)
-        next = self.request.GET.get("next", "")
-        if not next and self.extra_context and "next" in self.extra_context:
-            next = self.extra_context.get("next", "")
-        return f"{url}?next={next}"
-
-
 class EditorialView(TemplateView):
     template_name = "editorial/base.html"
 
@@ -330,77 +304,6 @@ def vote(request, erp_slug):
                 request, messages.SUCCESS, "Votre vote a été enregistré."
             )
     return redirect(erp.get_absolute_url())
-
-
-@login_required
-def mon_compte(request):
-    return render(request, "compte/index.html")
-
-
-@login_required
-def mon_identifiant(request):
-    if request.method == "POST":
-        form = forms.UsernameChangeForm(request.POST)
-        if form.is_valid():
-            username = form.cleaned_data["username"]
-            user = get_user_model().objects.get(id=request.user.id)
-            old_username = user.username
-            user.username = username
-            user.save()
-            LogEntry.objects.log_action(
-                user_id=request.user.id,
-                content_type_id=ContentType.objects.get_for_model(user).pk,
-                object_id=user.id,
-                object_repr=username,
-                action_flag=CHANGE,
-                change_message=f"Changement de nom d'utilisateur (avant: {old_username})",
-            )
-            messages.add_message(
-                request,
-                messages.SUCCESS,
-                f"Votre nom d'utilisateur a été changé en {user.username}.",
-            )
-            return redirect("mon_identifiant")
-    else:
-        form = forms.UsernameChangeForm(initial={"username": request.user.username})
-    return render(
-        request,
-        "compte/mon_identifiant.html",
-        context={"form": form},
-    )
-
-
-@login_required
-def mon_email(request):
-    if request.method == "POST":
-        form = forms.EmailChangeForm(request.POST)
-        if form.is_valid():
-            email = form.cleaned_data["email"]
-            user = get_user_model().objects.get(id=request.user.id)
-            old_email = user.email
-            user.email = email
-            user.save()
-            LogEntry.objects.log_action(
-                user_id=request.user.id,
-                content_type_id=ContentType.objects.get_for_model(user).pk,
-                object_id=user.id,
-                object_repr=email,
-                action_flag=CHANGE,
-                change_message=f"Changement de nom d'utilisateur (avant: {old_email})",
-            )
-            messages.add_message(
-                request,
-                messages.SUCCESS,
-                f"Votre email a été changé en {user.email}.",
-            )
-            return redirect("mon_identifiant")
-    else:
-        form = forms.UsernameChangeForm(initial={"email": request.user.email})
-    return render(
-        request,
-        "compte/mon_email.html",
-        context={"form": form},
-    )
 
 
 @login_required
